lite förbättringar

+++ deleteAcc.php
@@ -14,7 +14,14 @@ $app->getDB()->query($query);
 $query = "DELETE FROM comments WHERE userId = $id";
 $app->getDB()->query($query);
 
-unlink("images/" . $image);
+$query = "SELECT * FROM `users` WHERE image = '$image'";
+$result = $app->getDB()->query($query);
+
+if($result->num_rows == 1)
+{   
+    unlink("images/" . $image);
+}
+
 $query = "DELETE FROM users WHERE id = $id";
 $app->getDB()->query($query);
